requirements_macos_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blas=1.0=openblas
brotlipy=0.7.0=py311h80987f9_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=h80987f9_0
ca-certificates=2023.08.22=hca03da5_0
certifi=2023.7.22=py311hca03da5_0
cffi=1.15.1=py311h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=39.0.1=py311h834c97f_2
cython=0.29.35=py311h313beb8_0
filelock=3.9.0=py311hca03da5_0
freetype=2.12.1=h1192e45_0
fvcore=0.1.5.post20221221=pypi_0
giflib=5.2.1=h80987f9_3
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py311h40f64dc_0
h5py=3.9.0=py311hba6ad2f_0
hdf5=1.12.1=h05c076b_3
idna=3.4=py311hca03da5_0
imageio=2.31.4=py311hca03da5_0
iopath=0.1.10=pypi_0
jinja2=3.1.2=py311hca03da5_0
jpeg=9e=h80987f9_1
krb5=1.20.1=hf3e1bf2_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libcurl=8.2.1=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libigl=2.4.1=pypi_0
libnghttp2=1.52.0=h62f6fdd_1
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libssh2=1.10.0=h02f6b3c_2
libtiff=4.5.0=h313beb8_2
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
line-profiler=4.1.1=pypi_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py311h80987f9_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.2.1=py311hca03da5_0
ncurses=6.4=h313beb8_0
networkx=2.8.4=py311hca03da5_1
numpy=1.25.1=pypi_0
numpy-base=1.25.0=py311hfbfe69c_0
openssl=3.0.12=h1a28f6b_0
pillow=9.4.0=py311h313beb8_0
pip=23.1.2=py311hca03da5_0
plyfile=1.0.1=pypi_0
portalocker=2.7.0=pypi_0
pybind11=2.10.4=py311h48ca7d4_0
pybind11-global=2.10.4=py311h48ca7d4_0
pycparser=2.21=pyhd3eb1b0_0
pyglet=1.5.27=pypi_0
pyopenssl=23.0.0=py311hca03da5_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.4=hb885b13_0
pytorch=2.0.1=py3.11_0
pytorch3d=0.7.4=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h1a28f6b_0
requests=2.29.0=py311hca03da5_0
scipy=1.11.1=pypi_0
setuptools=67.8.0=py311hca03da5_0
sqlite=3.41.2=h80987f9_0
sympy=1.11.1=py311hca03da5_0
tabulate=0.9.0=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.12=hb8d0fd4_0
torchaudio=2.0.2=py311_cpu
torchvision=0.15.2=cpu_py311he74fb5d_0
tqdm=4.66.1=pypi_0
trimesh=3.22.4=pypi_0
typing_extensions=4.6.3=py311hca03da5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311hca03da5_0
wheel=0.38.4=py311hca03da5_0
xz=5.4.2=h80987f9_0
yacs=0.1.8=pypi_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
